user_app/requirements.txt
flask
pyjwt
boto3
mysql-connector-python
pillow
requests
gunicorn
celery
redis
pytest
